frontend/requirements.txt
alembic~=1.13.1
colorama~=0.4.6
colorlog~=6.8.2
contourpy~=1.2.1
cycler~=0.12.1
fonttools~=4.51.0
greenlet~=3.0.3
joblib~=1.4.2
kiwisolver~=1.4.5
lightgbm~=4.3.0
mako~=1.3.3
markupsafe~=2.1.5
matplotlib~=3.8.4
numpy~=1.26.4
optuna~=3.6.1
packaging~=24.0
pandas~=2.2.2
pillow~=10.3.0
pyparsing~=3.1.2
pytz~=2024.1
pyyaml~=6.0.1
scikit-learn~=1.4.2
scipy~=1.13.0
seaborn~=0.13.2
setuptools~=60.2.0
six~=1.16.0
sqlalchemy~=2.0.30
threadpoolctl~=3.5.0
tqdm~=4.66.4
typing-extensions~=4.11.0
tzdata~=2024.1
wheel~=0.37.1
fastapi~=0.111.0
uvicorn~=0.30.0
pydantic~=2.7.2
streamlit~=1.35.0
plotly~=5.22.0
millify~=0.1.1